Minor renaming
* Don't reference "Humbug" inside the humbug source tree.
